feat(v0.0.6): stdlib + 5 language gaps resolved (G1/G2/G3/G4/G7)
std/math — 22 native functions (trig, hyperbolic, exp/log, round, abs/max/min, pow) + PI and E constants
std/random — xoshiro256++ generator: entero(a,b), rango(n), peso_f64()
G1 — canonical deep update arr[i>j]$~ val (parser + interpreter, recursive deep_update_value)
G2 — $~ on named tuples: by 1-based position and by field name string
G4 — TypeConstraint::Numeric defaults to Float; Int→Float compatible, Int-evidence paths still infer Exact(Int)
G7 — named functions capture module_aliases at definition; restored before HOF call regardless of call depth

Drives Zofia Phases 1–5: all math gaps closed, all $~ update patterns supported.

+++ crates/zymbol-semantic/src/type_check.rs
@@ -130,7 +130,7 @@ impl TypeConstraint {
     fn to_type(&self) -> ZymbolType {
         match self {
             TypeConstraint::Exact(t) => t.clone(),
-            TypeConstraint::Numeric => ZymbolType::Int, // Default to Int if only numeric constraint
+            TypeConstraint::Numeric => ZymbolType::Float, // Numeric-only params accept both Int and Float (Int→Float is compatible)
             TypeConstraint::Boolean => ZymbolType::Bool,
             TypeConstraint::CompatibleWith(t) => t.clone(),
             TypeConstraint::Unconstrained => ZymbolType::Any,
